Generalize set-matrix to have index limits
Index limits may be passed to set-matrix and to some of the functions
dependent on it, so that e.g. a tridiagonal matrix can be defined
without iterating over all elements.  If the optional argument
'pass-element is non-NIL, the function will be passed a final argument
which is the prior value of the current element.