Fix defmfun expansion of default optional/key values
For methods on marrays, all default values for optional and key
arguments were wrongly coerced to the array element type.  This
doesn't affect any current definitions in GSLL, but FFT functions with
the stride specified, it will wrongly change the default value to the
element type.  This has been fixed by changing #'actual-class-arglist
so that it looks up the type intended to be passed to the C function
and coerces to that if it is available.  Otherwise the expansion
leaves the default value alone.