Unify forward FFT into new file and function
The function #'forward-fourier-transform selects among the 12 GSL
forward FFT functions, depending on whether the length of the vector
is a power of 2, whether the :half-complex flag is set, and the type
of the vector.  This exported function, along with its supporting
internal functions, are in the file forward.lisp.  The dimension
argument for fft-pulse-test is now mandatory.  All mobjects have been
renamed so that they end with either -single-float or -double-float,
consistent with the naming scheme in the rest of GSLL.  Eliminated
(evidently) obsolete example file fft-example.lisp.