Default wavetable/workspace; remove unused constant in GSL 1.13; fix dependency
In FFT, the default wavetable and workspace arguments to functions now
all use #'make-fft-wavetable and make-fft-workspace.  Removed
definition of +cgsm-gauss+ because evidently GSL_CONST_CGSM_GAUSS is
no longer in GSL as of 1.13.  Added dependency on init for
physical-consants.  Trivial change to names of systems in asd files to
be more standard.