Vector size passed to GSL functions now take stride into account.
For the FFT functions in GSLL, the vector sizes cannot be supplied by the
caller, but the stride can. The vector size passed to GSL is not the actual
vector size, but rather the size of the vector taking the stride into account:
(floor (size vector) stride). This is now fixed.