Split fft-noise test into FFT and DFT; test generation with supplied answer
The FFT noise test test-fft-noise calls a new split-out function
forward-fft-rc as well as the DFT "answer".  In preparation for the
generation of the test files, several functions in generate-tests.lisp
have been modified to accept a known answer, in lieu of evaluating the
form and saving that as the reference.