Added some extra functions for use with FFTs
fft-frequency-vector can be used to make a vector that contains the actual
sample frequencies (so you don't have to guess or write your own).

fft-shift and fft-inverse-shift shift the zero frequency to the center (for
easy plotting), and the inverse (for proper inverse FFT).

These functions are comparable to fftfreq, fftshift and ifftshift in numpy (if
anyone cares...).