Split off grid directory
Eventually, the definitions used for dealing with types and foreign
arrays that aren't specific to GSL will be split off into another
system, so that it will be easier to build other foreign library
interfaces using these definitions.  As a start, the grid/ directory
defines a package 'c-array and some of the type definitions formerly
in init/.  Everything compiles and tests as before, on SBCL
TOTAL: 1627 assertions passed, 6 failed, 0 execution errors.