Replace create-matrix and constant-matrix using grid functions
Replace create-matrix and constant-matrix by using the grid functions
grid:map-grid and grid:make-grid; eliminate create-vector because
vectors can be generated with create-matrix and using nil for the
second dimension argument.  In the grid:make-grid-data method for
make-marray, the first argument, which is the element type, needs to
be extracted from the rest spec with a newly-changed
grid:spec-scalar-p because compound types like (complex double-float)
were not being recognized as element types.