Remove grid to the gsd/c-array repository
The grid subdirectory is now the c-array system within the GSD
repository, as it has no GSL-dependent definitions and is potentially
useful on its own for other C libraries.  There is a bare minimum
mention of GSD in the documentation, which needs to be enhanced as it
is required for GSLL to compile/load.