Use CL RNG object, and not the foreign pointer in simulated annealing
Jame Wright pointed out that the function trivial-example-step tried
to apply the function #'sample to what turns out to be a foreign
pointer, but it should take a CL object.  This is because we are
making a random number generator, passing the foreign pointer to GSL,
and then it is passing that pointer back to our routines so that we
can use it in the step function.  This has been changed so that there
is a special variable cl-generator which is bound to the CL object in
#'simulated-annealing, and then used in the step function, while the
foreign pointer being returned to us is ignored.