Correct return values from jacobian-elliptic-functions; add tests
The GSL function gsl_sf_elljac_e writes sn, cn, dn values to doubles
for which the pointer has been passed in, as shown in
/usr/include/gsl/gsl_sf_elljac.h, and not to sf-result as one might be
led to conclude from the "_e" at the end of the name.  This is now
fixed in GSLL, so that the tests, which sometimes passed and sometimes
didn't depending on what was run before it, now should always work.
The single non-error test been replaced by the 35 GSL tests in
gsl/specfunc/test_sf.c, and they all pass in SBCL Linux AMD64.