C float parser is matching to nothing
The cl-ppcre C float parser is matching to nothing because there's no
single part of a float that is required; d. .d and d all possible, so
none of digit-before-decimal, decimal, or digit-after-decimal are
required.  Workaround is to check for length of 'mant and just return
the string if it's zero.