Fix assert-to-tolerance; remaining exponential-function tests ported
The macro #'assert-to-tolerance was not correct but now maps to
sf-check-results.  The order of arguments in the sf-frac-diff call in
sf-check-single has reversed to match test_sf_check_result in the GSL
code.  The remaining exponential-function tests have been ported
except for the one involving gsl_sf_exprel_n_CF_e which hasn't been
ported.  All the tests except one pass.