Fix fsbv:defcfun call
The expansion of a foreign structure-by-value function did not
substitute the foreign function name.  It turns out this didn't matter
because cffi::parse-name-and-options used by fsbv:defcfun simply
ignored the symbol 'GSL-NAME for the foreign function name and pulled
the actual name out of the symbol-name for the Lisp symbol of the
function name.  This fix makes it clearer in the macroexpansion
however.