Arrays passed to and received from callbacks are foreign-arrays
Using the new function faify-form, pointers passed by GSL to callbacks
are converted to foreign-array objects if scalarsp is nil.  This has
been tested on multidimensional minimization only so far, and works.
Note that this is potentially inefficient, as each time the callback
is called, new objects are made and in the callback itself any grefs
involve a generic function dispatch.  However, it is easier to write
the callback than dealing with the raw pointer.