Fix for random errors in tests
The test functions did not properly initialize newly created vectors to
contain unique values. This caused comparison of elements that are out of the
stride to randomly succeed. The new test functions properly initialize the
vectors just like is done in the GSL test functions. The result is that
comparison of elements that are out of stride should fail consistently.

This automatically means that the comparison should only check elements which
are on stride, just like is done in GSL. This is not fixed yet, so the
fast-fourier-transform unit tests are still left disabled for now.