Added non-radix-2 tests for radix-2-sized vectors
Effectively, non-radix-2 algorithm tests on power-of-2 vector sizes are added.
The result is, that some tests fail, and the cause at this point is not yet
clear. The tests that fail, are all related to non-radix-2 algorithms working
on power-of-2 vector sizes.

Only tests that are missing, are the so-called bitreverse tests.