New macro access-value
New macro access-value to call gsl_*_set and gsl_*_get, as yet unused.
It will eventually replace get-value and (setf get-value) as used in
reference-foreign-element, but has the arguments in the right order
and because it's a macro, expands immediately to a
cffi:foreign-funcall instead of doing a generic function dispatch on
an eql spcializer of the class name.