New macros #'maref and #'(setf maref)
New macros #'maref and #'(setf maref) get and set elements of the GSL
array directly from the mpointer.  These are not used by users, but in
the make-funcallable-form expansion of callbacks when scalarsp = T is
specified.  This fixes an error in the argument order in the (setf
get-value) generic function that this replaces.