Shadow symbols from antik
With the export of unit and dimension names, there are conflicts with
gsl:psi, gsl:acceleration, gsl:knots.  There is no actual conflict in
the usage because Antik physical quantities only needs to use these
symbols as symbols, there is no binding.  So they are now
shadowing-imported into gsll.  

Added explicit "d0" to numbers in polynomial to insure they are read
as double-floats.