In actual-array-class, don't replace real class names
In actual-array-class, if 'category is not one of 'vector, 'matrix or
'both, then assume it's actually a class name and just return it
without trying to make a class name.  This fixes a problem with
e.g. elt+ where the second argument specializer is a class name of the
scalar.