Fix sf-result* definitions, eliminate ffexpand, fsbv:object
Fix sf-result and sf-result-e10 translation methods, and use in
complex-with-error and values-with-errors.  Some usage in gamma.lisp
updated, but there will be a need to replace sf-result with (:struct
sf-result) in all declarations for many special functions.

I removed ffexpand and everything it depended on because there is no
reason to make a side defcfun for call-by-value functions, and if FSBV
is missing CFFI already emits an error.  All usage of fsbv:object and
any thing from the FSBV package is now gone.