Substitute for dim0 and dim1 in funcallables when scalarsp=NIL
Previously, the funcallable form made by #'make-funcallable-form (for
e.g. ODE solvers) when scalarsp=T involved substituting the symbols
'dim0 and 'dim1, but when scalarsp=NIL, would not do the substitution.
This caused an error when those unevaluated and unbound symbols
appeared in the lambda.  By altering #'faify-form with an additional
'dimension-values argument and calling 
(value-from-dimensions argspec dimension-values) 
instead of 
(parse-callback-argspec argspec 'dimensions), 
this problem is fixed.