Add direct foreign-vector view of 1D histograms
Defined two functions, view-range-as-foreign-array and view-bin-as-foreign-array that proved direct foreign-vector view of 1D histograms (shared data memory with the foreign histogram). This maybe be a precursor to some eventual generic "view" function, so the function symbols are not exported yet as the name may change. Also, the defcstruct should be done as a grovel on the C header. However, the functions do work now.