Add Monte Carlo parameter setting
With GSL 1.13, the GSL API defines a way to set the parameters for the
MISER and VEGAS Monte Carlo methods. This commit defines two generic
functions, #'parameter and #'(setf parameter), to get and set the
parameters, respectively.  It loads but has not been tested.

The previous way of setting the parameters, with macros
> #'miser-parameter and #'vegas-parameter, are conditional to
> #+obsolete-gsl and so do not normally load.