Start API changes for GSL 2.0
Release 2.0 of the GSL library changed the API for many functions, and
this means that the previous versions of GSLL will not load for this
release of the library. This is a start at handling these changes, with
an initial focus on permitting GSLL to load and not yet insuring that
the changed defintions function correctly.

There are two such updates made here:
1) The elimination of multiple arguments in gsl_multifit_linear_svd, and
consequently an internal change to #'linear-mfit-svd.
2) The elimination of the argument n in gsl_sf_ellint_D_e and
consequently in #'elliptic-integral-D.

This report is helpful in identifying the GSL API changes:
http://abi-laboratory.pro/tracker/compat_report/gsl/1.16/2.0/75ebf/abi_compat_report.html