New argument to defmobject
Add key argument :initialize-when-making to defmobject, which controls
whether initialization (by calling reinitialize-instance or some other
function) happens when the make-* function defined by this macro is
called. Choices are
  T=yes always
  NIL=no
  :default-T=optional (default T)
  :default-NIL=optional (default NIL)
and the default is T to be consistent with previous behavior.

This is used in the newly defined defmobject 'permutation.