GSL2 use new Jacobian function gsl_multifit_fdfsolver_jac
Function #'jacobian calls GSL2 function gsl_multifit_fdfsolver_jac
with a second optional argument, this argument is the matrix into
which the jacobian is written. If it is NIL, than the matrix is
created with the new function #'make-jacobian-matrix. Covariance
calculation #'ls-covariance calls #'jacoabian as in GSL1, but needs to
pass result to #'mpointer to get a C pointer. To avoid reallocating
covariance matrix repeatedly, call #'make-jacobian-matrix outside the
iteration loop and pass as argument to ls-covariance, as shown in
example #'nonlinear-least-squares-example.